Move pytest plugin into a private module and co-locate all documentation
