+++ docs/source/library_reference.rst
@@ -17,95 +17,6 @@ Library Reference
 
 .. module:: cocotb
 
-.. _api-pytest-plugin:
-
-Pytest Plugin
-=============
-
-
-.. _api-pytest-plugin-fixtures:
-
-Fixtures
---------
-
-.. module:: cocotb_tools.pytest.plugin
-
-.. autofixture:: dut
-
-.. autofixture:: hdl_session
-
-.. autofixture:: hdl
-
-
-.. _api-pytest-plugin-markers:
-
-Markers
--------
-
-.. module:: cocotb_tools.pytest.mark
-
-.. autodecorator:: cocotb_runner
-
-.. autodecorator:: cocotb_test
-
-.. autodecorator:: cocotb_timeout
-
-.. autodecorator:: cocotb_library
-
-.. autodecorator:: cocotb_sources
-
-.. autodecorator:: cocotb_defines
-
-.. autodecorator:: cocotb_includes
-
-.. autodecorator:: cocotb_parameters
-
-.. autodecorator:: cocotb_plusargs
-
-.. autodecorator:: cocotb_env
-
-.. autodecorator:: cocotb_seed
-
-.. autodecorator:: cocotb_timescale
-
-.. autodecorator:: cocotb_always
-
-.. autodecorator:: cocotb_clean
-
-.. autodecorator:: cocotb_waves
-
-.. autodecorator:: cocotb_build_args
-
-.. autodecorator:: cocotb_elab_args
-
-.. autodecorator:: cocotb_test_args
-
-.. autodecorator:: cocotb_pre_cmd
-
-.. _api-pytest-plugin-hdl:
-
-
-HDL Fixture Request
--------------------
-
-.. module:: cocotb_tools.pytest.hdl
-
-.. autoclass:: HDL
-    :members:
-
-
-.. _api-pytest-plugin-hook-specs:
-
-Hook Specifications
--------------------
-
-.. module:: cocotb_tools.pytest.hookspecs
-
-.. autofunction:: pytest_cocotb_make_hdl
-
-.. autofunction:: pytest_cocotb_make_runner
-
-
 .. _api-runner:
 
 Python Test Runner
